evict name maper from the TXT registry code
